# swagger.yaml
openapi: 3.0.1
info:
  title: "API Lịch Giảng Dạy - Đại học Thủy Lợi"
  description: "Tài liệu đặc tả API toàn diện cho ứng dụng Lịch Giảng Dạy. Phiên bản này bao gồm tất cả các tài nguyên, DTO, và logic duyệt 2 cấp mới nhất."
  version: "2.0.0"
servers:
  - url: http://localhost:8080
    description: Máy chủ phát triển local

# ================================================
# 1. Security Definitions
# ================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ================================================
  # 2. Reusable Schemas (DTOs)
  # ================================================
  schemas:
...
    ScheduleStatus: { type: string, description: "Trạng thái của một buổi học", enum: [NOT_TAUGHT, TAUGHT, ABSENT_APPROVED, ABSENT_UNAPPROVED, MAKEUP_TAUGHT] }
...
    LecturerDTO: { type: object, properties: { lecturerId: { type: integer, readOnly: true }, lecturerCode: { type: string }, fullName: { type: string }, email: { type: string, format: email }, department: { $ref: '#/components/schemas/DepartmentDTO' } } }
...
    StudentDTO: { type: object, properties: { id: { type: integer, readOnly: true }, studentCode: { type: string }, fullName: { type: string }, studentClass: { $ref: '#/components/schemas/StudentClassDTO' } } }
...
    DepartmentDTO: { type: object, properties: { departmentId: { type: integer, readOnly: true }, departmentName: { type: string } } }
    SubjectDTO: { type: object, properties: { subjectId: { type: integer, readOnly: true }, subjectCode: { type: string }, subjectName: { type: string } } }
    StudentClassDTO: { type: object, properties: { id: { type: integer, readOnly: true }, className: { type: string } } }
...
    AssignmentDTO: { type: object, properties: { assignmentId: { type: integer, readOnly: true }, lecturer: { $ref: '#/components/schemas/LecturerDTO' }, subject: { $ref: '#/components/schemas/SubjectDTO' }, studentClass: { $ref: '#/components/schemas/StudentClassDTO' }, theorySession: { type: integer }, practiceSession: { type: integer } } }
...
    ScheduleDTO: { type: object, properties: { sessionId: { type: integer, readOnly: true }, sessionDate: { type: string, format: date-time }, startPeriod: { type: integer }, endPeriod: { type: integer }, classroom: { type: string }, content: { type: string, nullable: true }, notes: { type: string, nullable: true }, status: { $ref: '#/components/schemas/ScheduleStatus' }, assignment: { $ref: '#/components/schemas/AssignmentDTO' } } }
    CreateScheduleDTO: { type: object, properties: { assignmentId: { type: integer }, sessionDate: { type: string, format: date-time }, startPeriod: { type: integer }, endPeriod: { type: integer }, classroom: { type: string } } }
...
    AttendanceDTO: { type: object, properties: { id: { type: integer, readOnly: true }, student: { $ref: '#/components/schemas/StudentDTO' }, isPresent: { type: boolean }, timestamp: { type: string, format: date-time, nullable: true } } }
    TakeAttendanceDTO: { type: object, properties: { studentId: { type: integer }, isPresent: { type: boolean } } }
    TakeAttendanceRequestDTO: { type: object, description: "Danh sách điểm danh cho một buổi học", properties: { attendances: { type: array, items: { $ref: '#/components/schemas/TakeAttendanceDTO' } } } }
...
security:
  - bearerAuth: []

# ================================================
# 4. API Endpoint Definitions
# ================================================
...
paths:
...
  /api/schedules:
    get: { tags: [Schedule], summary: "[Admin] Lấy tất cả lịch trình", responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ScheduleDTO' } } } } } } }
    post: { tags: [Schedule], summary: "[Admin] Tạo một lịch trình mới", requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateScheduleDTO' } } } }, responses: { '201': { description: Created } } }
  /api/schedules/{id}:
    get: { tags: [Schedule], summary: "Lấy lịch trình theo ID", parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ScheduleDTO' } } } } } }
    put: { tags: [Schedule], summary: "[Admin] Cập nhật lịch trình", parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateScheduleDTO' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ScheduleDTO' } } } } } }
    delete: { tags: [Schedule], summary: "[Admin] Xóa lịch trình", parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '204': { description: "Xóa thành công" } } }
...
  /api/schedules/lecturer/{email}:
    get: { tags: [Schedule], summary: "Lấy lịch trình của giảng viên theo email", parameters: [{ name: email, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ScheduleDTO' } } } } } } }
  /api/attendance/schedule/{scheduleId}:
    get: { tags: [Attendance], summary: "Lấy danh sách điểm danh của một buổi học", parameters: [{ name: scheduleId, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AttendanceDTO' } } } } } } }
    post: { tags: [Attendance], summary: "[Lecturer] Lưu thông tin điểm danh cho buổi học", parameters: [{ name: scheduleId, in: path, required: true, schema: { type: integer } }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TakeAttendanceRequestDTO' } } } }, responses: { '200': { description: OK } } }
...